requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
python-dotenv>=1.0.0
lxml>=4.9.0
line-bot-sdk>=3.0.0

